% Auto-generated: do not edit by hand
\name{testInput}

\alias{testInput}

\title{TestInput component}

\description{
Required description.
}

\usage{
testInput(placeholder=NULL)
}

\arguments{
\item{placeholder}{Character. Required description for each prop.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

